+++ app/Http/Livewire/Template/OneWay.php
@@ -3,12 +3,14 @@
 namespace App\Http\Livewire\Template;
 
 use App\Models\Template;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 use Illuminate\Support\Str;
 
 class OneWay extends Component
 {
+    use AuthorizesRequests;
     public $modalActionVisible = false;
     public $way = 1;
     public $type;
@@ -23,6 +25,7 @@ class OneWay extends Component
 
     public function create()
     {
+
         $this->validate();
 
         Template::create($this->modelData());
@@ -34,9 +37,11 @@ class OneWay extends Component
 
     public function update()
     {
+
         $this->validate();
 
         $template = Template::findOrFail($this->selectedItemId);
+        $this->authorize('UPDATE_CONTENT_USR', $template->user_id);
         $template->update($this->modelData());
 
         $this->modalActionVisible = false;
@@ -47,6 +52,7 @@ class OneWay extends Component
     public function edit($id)
     {
         $template = Template::findOrFail($id);
+        $this->authorize('UPDATE_CONTENT_USR', $template->user_id);
         $this->selectedItemId = $id;
         $this->type = $template->type;
         $this->name = $template->name;
@@ -59,6 +65,7 @@ class OneWay extends Component
     public function delete($id)
     {
         $template = Template::findOrFail($id);
+        $this->authorize('DELETE_CONTENT_USR', $template->user_id);
         $template->delete();
 
         $this->emit('refreshLivewireDatatable');
@@ -67,12 +74,12 @@ class OneWay extends Component
     public function modelData()
     {
         return [
-            'uuid'          => Str::uuid(),
-            'type'          => $this->way == 2 ? $this->type : 'text',
-            'name'          => $this->name,
-            'resource'      => $this->way,
-            'description'   => $this->description,
-            'user_id'       => Auth::user()->id,
+            'uuid' => Str::uuid(),
+            'type' => $this->way == 2 ? $this->type : 'text',
+            'name' => $this->name,
+            'resource' => $this->way,
+            'description' => $this->description,
+            'user_id' => Auth::user()->id,
         ];
     }
